+++ tests/test_player.py
@@ -1,8 +1,10 @@
 import unittest
+from unittest.mock import patch
 
 from game.player import Player
 from game.tile import Tile
 from game.bagtiles import BagTiles
+from game.scrabble import ScrabbleGame
 
 
 class TestPlayer(unittest.TestCase):
@@ -10,12 +12,13 @@ class TestPlayer(unittest.TestCase):
     def test_init(self):
 
         player_1 = Player()
+        
 
         self.assertEqual(
 
             len(player_1.tiles),
 
-            0,
+            7,
 
         )
 
@@ -31,98 +34,76 @@ class TestPlayer(unittest.TestCase):
         player.get_tiles(3,bag1)
         self.assertEqual(len(player.rack),3)
 
+    
+    @patch('builtins.input', return_value='tobias')
+    def test_nickname (self, input_patched):
+        player = Player()
+        player.set_nickname()        
+        self.assertEqual(player.nickname, 'tobias')
+
+
+    def test_increase_score(self):
+            player_1=Player()
+            player_1.increse_score(2)
+            self.assertEqual(player_1.score,2)
 
 
 
-        # 'Test que el profe hizo'
-    # def test_validate_user_has_letters(self):
+    def test_validate_tiles(self):
+        location = (7,7)
+        orientation = 'H'
+        scrabbleGame = ScrabbleGame(players_count=3)
+        scrabbleGame.players[scrabbleGame.current_player].tiles = [Tile('C', 1), Tile('A', 1), Tile('S', 1), Tile('A', 1)]
+        word = 'CASA'
+        self.assertEqual(scrabbleGame.players[scrabbleGame.current_player].validate_tiles(word, location, orientation, scrabbleGame), True)
 
-    #     bag_tile = BagTiles()
 
-    #     bag_tile.tiles = [
+    def test_validate_tiles_wrong(self):
+        location = (7,7)
+        orientation = 'H'
+        scrabbleGame = ScrabbleGame(players_count=3)
+        scrabbleGame.players[scrabbleGame.current_player].tiles = []
+        word = 'CASA'
+        self.assertEqual(scrabbleGame.players[scrabbleGame.current_player].validate_tiles(word, location, orientation, scrabbleGame), False)
 
-    #         Tile(letter='H', value=1),
 
-    #         Tile(letter='O', value=1),
+    def test_remove_tiles(self):
+        player = Player()
+        player.tiles = [Tile('A', 1),Tile('B', 3),Tile('C', 3),Tile('D', 2),Tile('E', 1)]
+        player.remove_tiles('BAD')
+        self.assertEqual(len(player.tiles), 2)
+        self.assertEqual(player.tiles[0].letter, 'C')
+        self.assertEqual(player.tiles[1].letter, 'E')   
 
-    #         Tile(letter='L', value=1),
 
-    #         Tile(letter='A', value=1),
-
-    #         Tile(letter='C', value=1),
-
-    #         Tile(letter='U', value=1),
-
-    #         Tile(letter='M', value=1),
-
-    #     ]
-
-    #     player = Player(bag_tile)
-
-    #     tiles = [
-
-    #         Tile(letter='H', value=1),
-
-    #         Tile(letter='O', value=1),
-
-    #         Tile(letter='L', value=1),
-
-    #         Tile(letter='A', value=1),
-
-    #     ]
+    def test_remove_tiles_not_found(self):
+        player = Player()
+        player.tiles = [Tile('A', 1), Tile('B', 3), Tile('C', 3), Tile('D', 2), Tile('E', 1)]
+        player.remove_tiles('XYZ')
+        self.assertEqual(len(player.tiles), 5)
 
 
 
-    #     is_valid = player.has_letters(tiles)
- 
-
-    #     self.assertEqual(is_valid, True)
-
-
-
-    # def test_validate_fail_when_user_has_not_letters(self):
-
-    #     bag_tile = BagTiles()
-
-    #     bag_tile.tiles = [
-
-    #         Tile(letter='P', value=1),
-
-    #         Tile(letter='O', value=1),
-
-    #         Tile(letter='L', value=1),
-
-    #         Tile(letter='A', value=1),
-
-    #         Tile(letter='C', value=1),
-
-    #         Tile(letter='U', value=1),
-
-    #         Tile(letter='M', value=1),
-
-    #     ]
-
-    #     player = Player(bag_tile)
-
-    #     tiles = [
-
-    #         Tile(letter='H', value=1),
-
-    #         Tile(letter='O', value=1),
-
-    #         Tile(letter='L', value=1),
-
-    #         Tile(letter='A', value=1),
-
-    #     ]
-
-
-    #     is_valid = player.has_letters(tiles)
+    def test_get_connected_tile(self):
+        scrabbleGame = ScrabbleGame(players_count=3)
+        scrabbleGame.board.grid[7][7].add_letter(Tile('C', 1),scrabbleGame.board)
+        word = 'CASA'
+        location = (7,7)
+        orientation = 'H'
+        self.assertEqual(scrabbleGame.players[scrabbleGame.current_player].get_connected_tile(word, location, orientation, scrabbleGame), ['C'])
 
 
 
-    #     self.assertEqual(is_valid, False)
-
+    def test_get_connected_tile2(self):
+        scrabbleGame = ScrabbleGame(players_count=3)
+        scrabbleGame.board.grid[7][7].add_letter(Tile('C', 1),scrabbleGame.board)
+        scrabbleGame.board.grid[7][8].add_letter(Tile('A', 1),scrabbleGame.board)
+        scrabbleGame.board.grid[7][9].add_letter(Tile('S', 1),scrabbleGame.board)
+        scrabbleGame.board.grid[7][10].add_letter(Tile('A', 1),scrabbleGame.board)
+        word = 'CASAS'
+        location = (7,7)
+        orientation = 'H'
+        self.assertEqual(scrabbleGame.players[scrabbleGame.current_player].get_connected_tile(word, location, orientation, scrabbleGame), ['C', 'A', 'S', 'A'])
 
 
 if __name__ == '__main__':